FIX: Big refactor, move common code in a new folder
