+++ trieTools/accountStorageExporter/main.go
@@ -7,41 +7,24 @@ import (
 	"fmt"
 	"io/fs"
 	"io/ioutil"
-	"math/big"
 	"os"
-	"path"
 	"path/filepath"
 
 	"github.com/ElrondNetwork/elrond-go-core/core"
 	"github.com/ElrondNetwork/elrond-go-core/core/check"
 	"github.com/ElrondNetwork/elrond-go-core/core/pubkeyConverter"
 	logger "github.com/ElrondNetwork/elrond-go-logger"
-	elrondFactory "github.com/ElrondNetwork/elrond-go/cmd/node/factory"
 	"github.com/ElrondNetwork/elrond-go/common"
-	commonDisabled "github.com/ElrondNetwork/elrond-go/common/disabled"
-	"github.com/ElrondNetwork/elrond-go/common/logging"
-	"github.com/ElrondNetwork/elrond-go/epochStart/notifier"
 	"github.com/ElrondNetwork/elrond-go/state"
-	stateFactory "github.com/ElrondNetwork/elrond-go/state/factory"
-	"github.com/ElrondNetwork/elrond-go/state/storagePruningManager/disabled"
-	dbRemoverDisabled "github.com/ElrondNetwork/elrond-go/storage/databaseremover/disabled"
-	"github.com/ElrondNetwork/elrond-go/storage/factory"
-	"github.com/ElrondNetwork/elrond-go/storage/pruning"
-	"github.com/ElrondNetwork/elrond-go/testscommon"
-	"github.com/ElrondNetwork/elrond-go/trie"
 	"github.com/ElrondNetwork/elrond-tools-go/trieTools/accountStorageExporter/config"
 	"github.com/ElrondNetwork/elrond-tools-go/trieTools/trieToolsCommon"
-	"github.com/ElrondNetwork/elrond-tools-go/trieTools/trieToolsCommon/components"
 	"github.com/urfave/cli"
 )
 
 const (
-	defaultLogsPath      = "logs"
-	logFilePrefix        = "account-storage-exporter"
-	maxTrieLevelInMemory = 5
-	rootHashLength       = 32
-	addressLength        = 32
-	maxDirs              = 100
+	logFilePrefix  = "account-storage-exporter"
+	rootHashLength = 32
+	addressLength  = 32
 )
 
 func main() {
@@ -73,7 +56,7 @@ func main() {
 func startProcess(c *cli.Context) error {
 	flagsConfig := getFlagsConfig(c)
 
-	_, errLogger := attachFileLogger(log, flagsConfig)
+	_, errLogger := trieToolsCommon.AttachFileLogger(log, logFilePrefix, flagsConfig.ContextFlagsConfig)
 	if errLogger != nil {
 		return errLogger
 	}
@@ -93,7 +76,7 @@ func startProcess(c *cli.Context) error {
 		return fmt.Errorf("wrong root hash length: expected %d, got %d", rootHashLength, len(rootHash))
 	}
 
-	maxDBValue, err := getMaxDBValue(filepath.Join(flagsConfig.WorkingDir, flagsConfig.DbDir))
+	maxDBValue, err := trieToolsCommon.GetMaxDBValue(filepath.Join(flagsConfig.WorkingDir, flagsConfig.DbDir), log)
 	if err != nil {
 		return err
 	}
@@ -103,103 +86,13 @@ func startProcess(c *cli.Context) error {
 	return exportStorage(flagsConfig.Address, flagsConfig, rootHash, maxDBValue)
 }
 
-func attachFileLogger(log logger.Logger, flagsConfig config.ContextFlagsConfigAddr) (elrondFactory.FileLoggingHandler, error) {
-	var fileLogging elrondFactory.FileLoggingHandler
-	var err error
-	if flagsConfig.SaveLogFile {
-		fileLogging, err = logging.NewFileLogging(logging.ArgsFileLogging{
-			WorkingDir:      flagsConfig.WorkingDir,
-			DefaultLogsPath: defaultLogsPath,
-			LogFilePrefix:   logFilePrefix,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("%w creating a log file", err)
-		}
-	}
-
-	err = logger.SetDisplayByteSlice(logger.ToHex)
-	log.LogIfError(err)
-	logger.ToggleLoggerName(flagsConfig.EnableLogName)
-	logLevelFlagValue := flagsConfig.LogLevel
-	err = logger.SetLogLevel(logLevelFlagValue)
-	if err != nil {
-		return nil, err
-	}
-
-	if flagsConfig.DisableAnsiColor {
-		err = logger.RemoveLogObserver(os.Stdout)
-		if err != nil {
-			return nil, err
-		}
-
-		err = logger.AddLogObserver(os.Stdout, &logger.PlainFormatter{})
-		if err != nil {
-			return nil, err
-		}
-	}
-	log.Trace("logger updated", "level", logLevelFlagValue, "disable ANSI color", flagsConfig.DisableAnsiColor)
-
-	return fileLogging, nil
-}
-
-func getMaxDBValue(parentDir string) (int, error) {
-	contents, err := ioutil.ReadDir(parentDir)
-	if err != nil {
-		return 0, err
-	}
-
-	directories := make([]string, 0)
-	for _, c := range contents {
-		if !c.IsDir() {
-			continue
-		}
-
-		_, ok := big.NewInt(0).SetString(c.Name(), 10)
-		if !ok {
-			log.Debug("DB directory found that will not be taken into account", "name", c.Name())
-			continue
-		}
-
-		directories = append(directories, c.Name())
-	}
-
-	numDirs := 0
-	for i := 0; i < maxDirs; i++ {
-		expectedDir := fmt.Sprintf("%d", i)
-		if !contains(directories, expectedDir) {
-			break
-		}
-
-		numDirs++
-	}
-
-	if numDirs == 0 {
-		return 0, fmt.Errorf("missing ordered directories in %s, like 0, 1 and so on", parentDir)
-	}
-	if numDirs != len(directories) {
-		return 0, fmt.Errorf("unordered directories in %s, like 0, 1 and so on", parentDir)
-	}
-
-	return numDirs - 1, nil
-}
-
-func contains(haystack []string, needle string) bool {
-	for _, h := range haystack {
-		if h == needle {
-			return true
-		}
-	}
-
-	return false
-}
-
 func exportStorage(address string, flags config.ContextFlagsConfigAddr, mainRootHash []byte, maxDBValue int) error {
 	addressConverter, err := pubkeyConverter.NewBech32PubkeyConverter(addressLength, log)
 	if err != nil {
 		return err
 	}
 
-	tr, err := getTrie(flags, maxDBValue)
+	tr, err := trieToolsCommon.GetTrie(flags.ContextFlagsConfig, maxDBValue)
 	if err != nil {
 		return err
 	}
@@ -209,7 +102,7 @@ func exportStorage(address string, flags config.ContextFlagsConfigAddr, mainRoot
 		log.LogIfError(errNotCritical)
 	}()
 
-	accDb, err := newAccountsAdapter(tr)
+	accDb, err := trieToolsCommon.NewAccountsAdapter(tr)
 	if err != nil {
 		return err
 	}
@@ -274,56 +167,4 @@ func exportStorage(address string, flags config.ContextFlagsConfigAddr, mainRoot
 	log.Info("key-value map", "value", keyValueMap)
 
 	return nil
-}
-
-func getTrie(flags config.ContextFlagsConfigAddr, maxDBValue int) (common.Trie, error) {
-	localDbConfig := dbConfig // copy
-	localDbConfig.FilePath = path.Join(flags.WorkingDir, flags.DbDir)
-
-	dbPath := path.Join(flags.WorkingDir, flags.DbDir)
-	args := &pruning.StorerArgs{
-		Identifier:                "",
-		ShardCoordinator:          testscommon.NewMultiShardsCoordinatorMock(1),
-		CacheConf:                 cacheConfig,
-		PathManager:               components.NewSimplePathManager(dbPath),
-		DbPath:                    "",
-		PersisterFactory:          factory.NewPersisterFactory(localDbConfig),
-		Notifier:                  notifier.NewManualEpochStartNotifier(),
-		OldDataCleanerProvider:    &testscommon.OldDataCleanerProviderStub{},
-		CustomDatabaseRemover:     dbRemoverDisabled.NewDisabledCustomDatabaseRemover(),
-		MaxBatchSize:              45000,
-		NumOfEpochsToKeep:         uint32(maxDBValue) + 1,
-		NumOfActivePersisters:     uint32(maxDBValue) + 1,
-		StartingEpoch:             uint32(maxDBValue),
-		PruningEnabled:            true,
-		EnabledDbLookupExtensions: false,
-	}
-
-	db, err := pruning.NewTriePruningStorer(args)
-	if err != nil {
-		return nil, err
-	}
-
-	tsm, err := trie.NewTrieStorageManagerWithoutPruning(db)
-	if err != nil {
-		return nil, err
-	}
-
-	return trie.NewTrie(tsm, trieToolsCommon.Marshaller, trieToolsCommon.Hasher, maxTrieLevelInMemory)
-}
-
-func newAccountsAdapter(trie common.Trie) (state.AccountsAdapter, error) {
-	accCreator := stateFactory.NewAccountCreator()
-	storagePruningManager := disabled.NewDisabledStoragePruningManager()
-	accountsAdapter, err := state.NewAccountsDB(state.ArgsAccountsDB{
-		Trie:                  trie,
-		Hasher:                trieToolsCommon.Hasher,
-		Marshaller:            trieToolsCommon.Marshaller,
-		AccountFactory:        accCreator,
-		StoragePruningManager: storagePruningManager,
-		ProcessingMode:        common.Normal,
-		ProcessStatusHandler:  commonDisabled.NewProcessStatusHandler(),
-	})
-
-	return accountsAdapter, err
 }
